+++ src/xlsxbook.h
@@ -50,6 +50,8 @@ class xlsxbook {
     Rcpp::List            character_formatted_; // data frame
     Rcpp::NumericVector   height_;          // Provided to cell constructor
     Rcpp::NumericVector   width_;           // Provided to cell constructor
+    Rcpp::NumericVector   rowOutlineLevel_; // Provided to cell constructor
+    Rcpp::NumericVector   colOutlineLevel_; // Provided to cell constructor
     Rcpp::CharacterVector style_format_;    // cellXfs xfId links to cellStyleXfs entry
     Rcpp::IntegerVector   local_format_id_; // cell 'c' links to cellXfs entry